Move SetPZ69DisplayBytesDefault into PZ69DisplayBytes function
